Fix out-of-bounds read in ShannonEntropyTracker for maximum symbol
ShannonEntropyTracker::UpdateSymbols grows |frequencies_| so that the
incoming symbol is a valid index:

  if (frequencies_.size() <= symbol) {
    frequencies_.resize(symbol + 1, 0);
  }

|symbol| is uint32_t, so |symbol| + 1 is evaluated in uint32_t and wraps
to 0 when the symbol is UINT32_MAX. resize(0) then shrinks the table to
empty, and the very next statement indexes it:

  int &frequency = frequencies_[symbol];

which reads and writes far out of bounds. The comparison on the line
above has the same problem in reverse: size_t and uint32_t compare fine
on LP64, but the arithmetic that follows does not.

Casting to size_t before both the comparison and the addition keeps the
value at its true magnitude. Reachable from the encoder: the tracker is
fed ConvertSignedIntToSymbol() of a prediction residual, and a residual
near INT32_MIN zigzags to UINT32_MAX. Found with AddressSanitizer while
fuzzing the encoder with randomly generated meshes at high quantization.

Note this makes the allocation honest rather than small - a symbol that
large asks for a 16 GB frequency table and will now fail to allocate
instead of silently corrupting memory. Sizing the table by the largest
symbol seen is a separate design question and is left alone here.

+++ src/draco/compression/entropy/shannon_entropy.cc
@@ -69,8 +69,11 @@ ShannonEntropyTracker::EntropyData ShannonEntropyTracker::UpdateSymbols(
   ret_data.num_values += num_symbols;
   for (int i = 0; i < num_symbols; ++i) {
     const uint32_t symbol = symbols[i];
-    if (frequencies_.size() <= symbol) {
-      frequencies_.resize(symbol + 1, 0);
+    // Note the casts: |symbol| + 1 is evaluated in uint32_t and wraps to 0 for
+    // the maximum symbol value, which resizes the table to empty and makes the
+    // indexing below read out of bounds.
+    if (frequencies_.size() <= static_cast<size_t>(symbol)) {
+      frequencies_.resize(static_cast<size_t>(symbol) + 1, 0);
     }
 
     // Update the entropy of the stream. Note that entropy of |N| values
